Commit of Branching Assignment
Committing creation of documentation for the Branching Statement Assignment per step 222 of the C# course.  This app takes in various user inputs about a package, calculates and compares the inputs, and determines either the cost to ship the package or if the package is too large or heavy to ship.