Commit String Assignment
Committing creating and submission of String Assignment per step 233 of the C# course.  For this assignment, I created user provided input to create strings and performed a casing statement as well as a concatenation.  I then created a stringbuilder object, prompted user input, and amended the object to reflect the user input.