Commit of Methods and Objects Assignment
Committing creation of app per step 292 in the C# course.  This app demonstrates the ability to create a class, add properties and a method, then a second class that inherits the first class properties and methods, and adds an additional property.  Then in the main program an object of the second class is created and utilizes a superclass method to print object properties to the screen.